Test calibration conversion function

+++ pid.h
@@ -7,9 +7,9 @@ extern short Kd;
 extern short pe; // proportional error
 extern short ie; // integral error
 extern short de; // derivative error
-extern short set_point;
-extern short error;
-extern short T_meas;
+extern float set_point;
+extern float error;
+extern float T_meas;
 
 // Subroutine to run basic pid control loop
 void pid_controller_loop(void);
